fix: rename eval_task module to avoid shadowing by function import
Rename eval_task.py to eval_task_handler.py so the module name
no longer collides with the eval_task function re-exported in
__init__.py. This was causing @patch targets in TestTracedHandler
to resolve to the function instead of the module, breaking all
5 TracedHandler tests with AttributeError.
